ci: install golang 1.14.2 in windows

+++ scripts/jenkins/windows/install-tools.bat
@@ -2,5 +2,12 @@
 :: This script install the required tools
 ::
 
-choco config set cacheLocation %WORKSPACE%
-choco install golang --version %GO_VERSION% -y --no-progress
+if "%GO_VERSION%" == "1.14.2" (
+    git clone https://github.com/v1v/chocolatey-packages .chocolatey-packages
+    cd chocolatey-packages\golang\1.14.2
+    choco pack
+    choco install golang -s .
+) else (
+    choco config set cacheLocation %WORKSPACE%
+    choco install golang --version %GO_VERSION% -y --no-progress
+)
